Add throttle time metrics for async write (#11936)
![Screenshot 2025-01-07 at 2 38
25 PM](https://github.com/user-attachments/assets/799bc726-b5c4-4e64-b38e-db30f69d9b4b)


The async write added in
https://github.com/NVIDIA/spark-rapids/pull/11730 supports throttling
based on the number in-flight bytes to write. Knowing how much time the
query spends on waiting in the throttle would be useful to understand
query performance. This PR adds these new metrics which are also seen in
the screenshot above:

- total throttle time: total throttle time in a task
- avg/min/max throttle time per async write: average/min/max throttle
time per async write request

The new metrics are added as debug metrics. To do so, The
`create*Metrics()` in `GpuExec` has been moved to `GpuMetric` so that
they can be used in other places than `GpuExec`. To put related things
together, `GpuMetric` and `MetricsLevel` have been moved to a new
`GpuMetrics.scala` file.

---------
